MDLA-Android-Release/src/core/android_vpn.py
"""Android VPN Service Integration"""
from kivy.utils import platform
from typing import Optional, Dict, Any

if platform == 'android':
    try:
        from jnius import autoclass, PythonJavaClass, java_method
        from android.permissions import request_permissions, Permission
        
        # Android classes
        VpnService = autoclass('android.net.VpnService')
        Intent = autoclass('android.content.Intent')
        PythonActivity = autoclass('org.kivy.android.PythonActivity')
        Context = autoclass('android.content.Context')
        
    except ImportError:
        print("Android classes not available")
        VpnService = None


class AndroidVPNManager:
    """Manages Android VPN service integration"""
    
    def __init__(self):
        self.vpn_service = None
        self.is_android = platform == 'android'
        self.activity = None
        
        if self.is_android and VpnService:
            self.activity = PythonActivity.mActivity
    
    def request_vpn_permission(self) -> bool:
        """Request VPN permission from user"""
        if not self.is_android or not VpnService:
            return True  # Desktop mode
        
        try:
            # Check if VPN permission is already granted
            intent = VpnService.prepare(self.activity)
            if intent is None:
                # Permission already granted
                return True
            else:
                # Need to request permission
                # This would typically show a system dialog
                print("VPN permission required - user must grant in system dialog")
                return False
        except Exception as e:
            print(f"Error requesting VPN permission: {e}")
            return False
    
    def request_storage_permissions(self):
        """Request storage permissions for config files"""
        if not self.is_android:
            return
        
        try:
            request_permissions([
                Permission.WRITE_EXTERNAL_STORAGE,
                Permission.READ_EXTERNAL_STORAGE,
                Permission.INTERNET,
                Permission.ACCESS_NETWORK_STATE,
                Permission.CHANGE_NETWORK_STATE
            ])
        except Exception as e:
            print(f"Error requesting permissions: {e}")
    
    def create_vpn_interface(self, config: Dict[str, Any]) -> bool:
        """Create VPN interface using Android VpnService"""
        if not self.is_android or not VpnService:
            print("VPN interface created (desktop mode)")
            return True
        
        try:
            # This is a simplified example
            # In a real implementation, you would:
            # 1. Create a VPN service class extending VpnService
            # 2. Configure the VPN interface (IP, DNS, routes)
            # 3. Start the VPN core process
            # 4. Route traffic through the VPN
            
            print("Creating Android VPN interface...")
            
            # Mock implementation for now
            return True
            
        except Exception as e:
            print(f"Error creating VPN interface: {e}")
            return False
    
    def destroy_vpn_interface(self) -> bool:
        """Destroy VPN interface"""
        if not self.is_android:
            print("VPN interface destroyed (desktop mode)")
            return True
        
        try:
            # Stop VPN service and clean up
            print("Destroying Android VPN interface...")
            return True
            
        except Exception as e:
            print(f"Error destroying VPN interface: {e}")
            return False
    
    def get_vpn_stats(self) -> Dict[str, int]:
        """Get VPN traffic statistics"""
        if not self.is_android:
            # Mock stats for desktop
            return {
                "tx_bytes": 0,
                "rx_bytes": 0,
                "tx_packets": 0,
                "rx_packets": 0
            }
        
        try:
            # Read from VPN interface statistics
            # This would typically read from /proc/net/dev or similar
            return {
                "tx_bytes": 0,
                "rx_bytes": 0,
                "tx_packets": 0,
                "rx_packets": 0
            }
        except Exception as e:
            print(f"Error getting VPN stats: {e}")
            return {
                "tx_bytes": 0,
                "rx_bytes": 0,
                "tx_packets": 0,
                "rx_packets": 0
            }


# Example VPN Service class (would need to be implemented in Java/Kotlin)
"""
public class MDLAVpnService extends VpnService {
    private ParcelFileDescriptor vpnInterface;
    private Thread vpnThread;
    
    @Override
    public int onStartCommand(Intent intent, int flags, int startId) {
        // Start VPN
        startVpn();
        return START_STICKY;
    }
    
    private void startVpn() {
        // Configure VPN interface
        Builder builder = new Builder();
        builder.setMtu(1500);
        builder.addAddress("10.0.0.2", 24);
        builder.addDnsServer("8.8.8.8");
        builder.addRoute("0.0.0.0", 0);
        
        // Create interface
        vpnInterface = builder.establish();
        
        // Start traffic handling thread
        vpnThread = new Thread(this::handleTraffic);
        vpnThread.start();
    }
    
    private void handleTraffic() {
        // Handle VPN traffic
        // This would typically involve:
        // 1. Reading packets from VPN interface
        // 2. Forwarding to proxy (Xray/Sing-Box)
        // 3. Receiving responses and writing back
    }
    
    @Override
    public void onDestroy() {
        // Clean up
        if (vpnInterface != null) {
            try {
                vpnInterface.close();
            } catch (IOException e) {
                // Handle error
            }
        }
        
        if (vpnThread != null) {
            vpnThread.interrupt();
        }
        
        super.onDestroy();
    }
}
"""